+++ Assets/Audio/Scores LOOPS.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c2a5d33433370734e9bf3e3cbb8cd816
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
