+++ Assets/Audio/Scores LOOPS/Scores LOOP A LONG.wav.meta
@@ -1,0 +1,21 @@
+fileFormatVersion: 2
+guid: 699d93f840f17445395ff9614e871269
+timeCreated: 1440433481
+licenseType: Store
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
