+++ Assets/Audio/Scores LOOPS/Scores LOOP A.wav.meta
@@ -1,0 +1,21 @@
+fileFormatVersion: 2
+guid: b2ea41f78e7014720977393dc8038543
+timeCreated: 1440433701
+licenseType: Store
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
