+++ Assets/Audio/Scores LOOPS/Scores LOOP C.wav.meta
@@ -1,0 +1,21 @@
+fileFormatVersion: 2
+guid: 7344e402e62d94bbab19374c91f4ec71
+timeCreated: 1440433512
+licenseType: Store
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
