+++ Assets/Audio/Scores LOOPS/Scores LOOP END HIT.wav.meta
@@ -1,0 +1,21 @@
+fileFormatVersion: 2
+guid: 327d6ebb7b0d8471099a11c144bcce0e
+timeCreated: 1440433389
+licenseType: Store
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
