--- Assets/Materials/L.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 4b61771618f62e243adf6a10386b761c
-timeCreated: 1536285541
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
