+++ Assets/Models/boy.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9140ffdb241198846b21e80815b59aef
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
