--- Assets/Models/no.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4b62c7012b78f8040bf378a91190c3d9
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
