--- Assets/Models/no.fbx.meta
@@ -1,176 +1,0 @@
-fileFormatVersion: 2
-guid: a5c25e525227ff84e90a39920e507767
-ModelImporter:
-  serializedVersion: 23
-  fileIDToRecycleName:
-    100000: "boy_3\u539F"
-    100002: "Cinema_4D_\u7F16\u8F91\u5668"
-    100004: //RootNode
-    100006: "\u5173\u8282"
-    100008: "\u5173\u8282_1"
-    100010: "\u5173\u8282_1_2"
-    100012: "\u5173\u8282_1_3"
-    100014: "\u5173\u8282_2"
-    100016: "\u5173\u8282_2_2"
-    100018: "\u5173\u8282_2_3"
-    100020: "\u5173\u8282_2_4"
-    100022: "\u5173\u8282_3"
-    100024: "\u5173\u8282_3_2"
-    100026: "\u5173\u8282_3_3"
-    100028: "\u5173\u8282_3_4"
-    100030: "\u5173\u8282_4"
-    100032: "\u5173\u8282_4_2"
-    100034: "\u5173\u8282_4_3"
-    100036: "\u5173\u8282_5"
-    100038: "\u5173\u8282_5_2"
-    100040: "\u5173\u8282_5_3"
-    100042: "\u5173\u8282_6"
-    100044: "\u5173\u8282_6_2"
-    100046: "\u5173\u8282_6_3"
-    100048: "\u5173\u8282_7"
-    100050: "\u5173\u8282_7_2"
-    100052: "\u6839\u5BF9\u8C61"
-    400000: "boy_3\u539F"
-    400002: "Cinema_4D_\u7F16\u8F91\u5668"
-    400004: //RootNode
-    400006: "\u5173\u8282"
-    400008: "\u5173\u8282_1"
-    400010: "\u5173\u8282_1_2"
-    400012: "\u5173\u8282_1_3"
-    400014: "\u5173\u8282_2"
-    400016: "\u5173\u8282_2_2"
-    400018: "\u5173\u8282_2_3"
-    400020: "\u5173\u8282_2_4"
-    400022: "\u5173\u8282_3"
-    400024: "\u5173\u8282_3_2"
-    400026: "\u5173\u8282_3_3"
-    400028: "\u5173\u8282_3_4"
-    400030: "\u5173\u8282_4"
-    400032: "\u5173\u8282_4_2"
-    400034: "\u5173\u8282_4_3"
-    400036: "\u5173\u8282_5"
-    400038: "\u5173\u8282_5_2"
-    400040: "\u5173\u8282_5_3"
-    400042: "\u5173\u8282_6"
-    400044: "\u5173\u8282_6_2"
-    400046: "\u5173\u8282_6_3"
-    400048: "\u5173\u8282_7"
-    400050: "\u5173\u8282_7_2"
-    400052: "\u6839\u5BF9\u8C61"
-    2000000: "Cinema_4D_\u7F16\u8F91\u5668"
-    2100000: "\u6750\u8D28.2"
-    2100002: "\u6750\u8D28.1"
-    4300000: "boy_3\u539F"
-    7400000: no
-    9500000: //RootNode
-    13700000: "boy_3\u539F"
-  externalObjects: {}
-  materials:
-    importMaterials: 1
-    materialName: 0
-    materialSearch: 1
-    materialLocation: 1
-  animations:
-    legacyGenerateAnimations: 4
-    bakeSimulation: 0
-    resampleCurves: 1
-    optimizeGameObjects: 0
-    motionNodeName: 
-    rigImportErrors: 
-    rigImportWarnings: 
-    animationImportErrors: 
-    animationImportWarnings: 
-    animationRetargetingWarnings: 
-    animationDoRetargetingWarnings: 0
-    importAnimatedCustomProperties: 0
-    importConstraints: 0
-    animationCompression: 1
-    animationRotationError: 0.5
-    animationPositionError: 0.5
-    animationScaleError: 0.5
-    animationWrapMode: 0
-    extraExposedTransformPaths: []
-    extraUserProperties: []
-    clipAnimations:
-    - serializedVersion: 16
-      name: no
-      takeName: CINEMA_4D___
-      firstFrame: 0
-      lastFrame: 80
-      wrapMode: 0
-      orientationOffsetY: 0
-      level: 0
-      cycleOffset: 0
-      loop: 0
-      hasAdditiveReferencePose: 0
-      loopTime: 1
-      loopBlend: 0
-      loopBlendOrientation: 0
-      loopBlendPositionY: 0
-      loopBlendPositionXZ: 0
-      keepOriginalOrientation: 0
-      keepOriginalPositionY: 1
-      keepOriginalPositionXZ: 0
-      heightFromFeet: 0
-      mirror: 0
-      bodyMask: 01000000010000000100000001000000010000000100000001000000010000000100000001000000010000000100000001000000
-      curves: []
-      events: []
-      transformMask: []
-      maskType: 3
-      maskSource: {instanceID: 0}
-      additiveReferencePoseFrame: 0
-    isReadable: 1
-  meshes:
-    lODScreenPercentages: []
-    globalScale: 1
-    meshCompression: 0
-    addColliders: 0
-    importVisibility: 1
-    importBlendShapes: 1
-    importCameras: 1
-    importLights: 1
-    swapUVChannels: 0
-    generateSecondaryUV: 0
-    useFileUnits: 1
-    optimizeMeshForGPU: 1
-    keepQuads: 0
-    weldVertices: 1
-    preserveHierarchy: 0
-    indexFormat: 0
-    secondaryUVAngleDistortion: 8
-    secondaryUVAreaDistortion: 15.000001
-    secondaryUVHardAngle: 88
-    secondaryUVPackMargin: 4
-    useFileScale: 1
-    previousCalculatedGlobalScale: 0.01
-    hasPreviousCalculatedGlobalScale: 1
-  tangentSpace:
-    normalSmoothAngle: 60
-    normalImportMode: 0
-    tangentImportMode: 3
-    normalCalculationMode: 4
-  importAnimation: 1
-  copyAvatar: 0
-  humanDescription:
-    serializedVersion: 2
-    human: []
-    skeleton: []
-    armTwist: 0.5
-    foreArmTwist: 0.5
-    upperLegTwist: 0.5
-    legTwist: 0.5
-    armStretch: 0.05
-    legStretch: 0.05
-    feetSpacing: 0
-    rootMotionBoneName: 
-    hasTranslationDoF: 0
-    hasExtraRoot: 0
-    skeletonHasParents: 1
-  lastHumanDescriptionAvatarSource: {instanceID: 0}
-  animationType: 2
-  humanoidOversampling: 1
-  additionalBone: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
