--- Assets/Models/yes.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3360bc6ded1e719469ae109e4e5b7ca5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
