--- Assets/Models/男孩.fbx.meta
@@ -1,147 +1,0 @@
-fileFormatVersion: 2
-guid: 6d11bd37531b8442bb2c4248d19a17d4
-ModelImporter:
-  serializedVersion: 23
-  fileIDToRecycleName:
-    100000: "boy_3\u539F"
-    100002: "Cinema_4D_\u7F16\u8F91\u5668"
-    100004: "\u5173\u8282"
-    100006: "\u5173\u8282_1"
-    100008: "\u5173\u8282_1_2"
-    100010: "\u5173\u8282_1_3"
-    100012: "\u5173\u8282_2"
-    100014: "\u5173\u8282_2_2"
-    100016: "\u5173\u8282_2_3"
-    100018: "\u5173\u8282_2_4"
-    100020: "\u5173\u8282_3"
-    100022: "\u5173\u8282_3_2"
-    100024: "\u5173\u8282_3_3"
-    100026: "\u5173\u8282_3_4"
-    100028: "\u5173\u8282_4"
-    100030: "\u5173\u8282_4_2"
-    100032: "\u5173\u8282_4_3"
-    100034: "\u5173\u8282_5"
-    100036: "\u5173\u8282_5_2"
-    100038: "\u5173\u8282_5_3"
-    100040: "\u5173\u8282_6"
-    100042: "\u5173\u8282_6_2"
-    100044: "\u5173\u8282_6_3"
-    100046: "\u5173\u8282_7"
-    100048: "\u5173\u8282_7_2"
-    100050: "\u6839\u5BF9\u8C61"
-    100052: //RootNode
-    400000: "boy_3\u539F"
-    400002: "Cinema_4D_\u7F16\u8F91\u5668"
-    400004: "\u5173\u8282"
-    400006: "\u5173\u8282_1"
-    400008: "\u5173\u8282_1_2"
-    400010: "\u5173\u8282_1_3"
-    400012: "\u5173\u8282_2"
-    400014: "\u5173\u8282_2_2"
-    400016: "\u5173\u8282_2_3"
-    400018: "\u5173\u8282_2_4"
-    400020: "\u5173\u8282_3"
-    400022: "\u5173\u8282_3_2"
-    400024: "\u5173\u8282_3_3"
-    400026: "\u5173\u8282_3_4"
-    400028: "\u5173\u8282_4"
-    400030: "\u5173\u8282_4_2"
-    400032: "\u5173\u8282_4_3"
-    400034: "\u5173\u8282_5"
-    400036: "\u5173\u8282_5_2"
-    400038: "\u5173\u8282_5_3"
-    400040: "\u5173\u8282_6"
-    400042: "\u5173\u8282_6_2"
-    400044: "\u5173\u8282_6_3"
-    400046: "\u5173\u8282_7"
-    400048: "\u5173\u8282_7_2"
-    400050: "\u6839\u5BF9\u8C61"
-    400052: //RootNode
-    2000000: "Cinema_4D_\u7F16\u8F91\u5668"
-    2100000: No Name
-    4300000: "boy_3\u539F"
-    7400000: CINEMA_4D___
-    9500000: //RootNode
-    13700000: "boy_3\u539F"
-  externalObjects: {}
-  materials:
-    importMaterials: 1
-    materialName: 0
-    materialSearch: 1
-    materialLocation: 1
-  animations:
-    legacyGenerateAnimations: 4
-    bakeSimulation: 0
-    resampleCurves: 1
-    optimizeGameObjects: 0
-    motionNodeName: 
-    rigImportErrors: 
-    rigImportWarnings: 
-    animationImportErrors: 
-    animationImportWarnings: 
-    animationRetargetingWarnings: 
-    animationDoRetargetingWarnings: 0
-    importAnimatedCustomProperties: 0
-    importConstraints: 0
-    animationCompression: 1
-    animationRotationError: 0.5
-    animationPositionError: 0.5
-    animationScaleError: 0.5
-    animationWrapMode: 0
-    extraExposedTransformPaths: []
-    extraUserProperties: []
-    clipAnimations: []
-    isReadable: 1
-  meshes:
-    lODScreenPercentages: []
-    globalScale: 1
-    meshCompression: 0
-    addColliders: 0
-    importVisibility: 1
-    importBlendShapes: 1
-    importCameras: 1
-    importLights: 1
-    swapUVChannels: 0
-    generateSecondaryUV: 0
-    useFileUnits: 1
-    optimizeMeshForGPU: 1
-    keepQuads: 0
-    weldVertices: 1
-    preserveHierarchy: 0
-    indexFormat: 0
-    secondaryUVAngleDistortion: 8
-    secondaryUVAreaDistortion: 15.000001
-    secondaryUVHardAngle: 88
-    secondaryUVPackMargin: 4
-    useFileScale: 1
-    previousCalculatedGlobalScale: 0.01
-    hasPreviousCalculatedGlobalScale: 1
-  tangentSpace:
-    normalSmoothAngle: 60
-    normalImportMode: 0
-    tangentImportMode: 3
-    normalCalculationMode: 4
-  importAnimation: 1
-  copyAvatar: 0
-  humanDescription:
-    serializedVersion: 2
-    human: []
-    skeleton: []
-    armTwist: 0.5
-    foreArmTwist: 0.5
-    upperLegTwist: 0.5
-    legTwist: 0.5
-    armStretch: 0.05
-    legStretch: 0.05
-    feetSpacing: 0
-    rootMotionBoneName: 
-    hasTranslationDoF: 0
-    hasExtraRoot: 0
-    skeletonHasParents: 1
-  lastHumanDescriptionAvatarSource: {instanceID: 0}
-  animationType: 2
-  humanoidOversampling: 1
-  additionalBone: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
