--- Assets/Scripts/Done_Mover.cs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 7f7850f24249e4056b2189de03ae584c
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
