--- Assets/Scripts/InputManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: daeae23c611ca7046ac87bf65f486b40
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
