--- Assets/Scripts/UIManager.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: acf832a06793f844f93669f932568f71
-timeCreated: 1547634020
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
